вынес метод анимации смены экранов в отдельный модуль
